Project 1 - Coba-Medpack/requirements.txt
langchain
langchain-google-genai
pillow
python-dotenv
langchain
langchain_google_genai
langgraph
pydantic
fastapi
uvicorn
python-multipart
Pillow
python-dotenv
pandas
numpy
qdrant-client
Scikit-learn
pydantic_settings